bug fixed, exp() --> expf()

+++ hestonBarrier/hestonBarrier.cpp
@@ -172,7 +172,7 @@ void heston::sampleSIM(RNG* mt_rng, data_t* call,data_t* put,int num_sims)
 						continue;
 					mt_rng[i].BOX_MULLER(&num1[i][s],&num2[i][s],pVols[i][s]);
 
-					stockPrice[i][s]*=exp(ratio3-pVols[i][s]*0.5f+num1[i][s]*vol.correlation+num2[i][s]*ratio2);
+					stockPrice[i][s]*=expf(ratio3-pVols[i][s]*0.5f+num1[i][s]*vol.correlation+num2[i][s]*ratio2);
 					if(stockPrice[i][s]<bData.downBarrier || stockPrice[i][s]>bData.upBarrier)
 					{
 						bBarrier[i][s]=false;
